fix(#3790): revert to BYTE_ORDER for time — my two signed changes were both wrong (roborev job 51)
Reverts e808790d9, 655617e3f and 24f2899f5. Site 1 is to_be_bytes() again and
time is back on the merged-read raw-byte fast path.

WHAT DECIDED IT: a FOURTH time-ordering site nobody had named.
  1 types/comparator/custom.rs   Custom("time")
  2 types.rs                     Value::PartialOrd                signed
  3 merge/read_assembly.rs       merged-read fast path
  4 data_writer/complex.rs:831 -> schema_helpers::compare_cell_paths (a.cmp(b))
                                  UNSIGNED -- THE ON-DISK WRITER

Cassandra orders time by BYTE_ORDER (pinned cassandra-5.0.8, TimeType.java:
super(ComparisonType.BYTE_ORDER)) -- unsigned. CQLite's byte-parity-guarded writer
(site 4) already does that, so it is correct and must not change. Critically it
RE-SORTS UNCONDITIONALLY before writing, so Value::PartialOrd's signed order is
overridden and NEVER REACHES DISK for collection elements.

Round 7's finding was "the comparator disagrees with the writer, which is signed".
I verified the signedness of THE WRONG WRITER -- site 2, not site 4 -- and reversed
on it, moving the comparator away from Cassandra AND from our own on-disk bytes.
Round 9 then led me to change site 3 the same way, so the merged read path
REORDERED what the writer had correctly written: worse than the starting state.

The causal chain is worth keeping: round 9's premise ("the fast path's equivalence
assumption is false") was true ONLY BECAUSE round 7 had already made the comparator
signed. With the comparator unsigned the two routes are equivalent again, which is
exactly what makes the fast path legitimate. One wrong turn manufactured the
evidence for the next.

NEW GUARD, with its RED CONDITIONS MEASURED rather than claimed --
tests/issue_3790_merged_read_time_order.rs (ungated; write-support is default, so
core-tests runs it):
  * comparator signed AND time off the fast path (what actually happened) -> both
    negative cases FAIL. Caught.
  * time off the fast path alone -> PASSES; with an unsigned comparator the two
    routes agree, which IS the soundness condition.
  * comparator signed alone -> PASSES; this path never consults the comparator.
    That half is guarded by the unit suite's
    time_byte_order_places_negative_nanos_above_zero.
The file declares that scope, because a test covering less than its name suggests
is worse than one that says so; the two files are co-required.

Site 2 remains signed: a real latent inconsistency that cannot affect on-disk
collection order, owned by #3920.

No fixture can settle any of this -- TimeSerializer validates
0..=86_399_999_999_999, so no Cassandra-written SSTable can hold a negative time.
The whole disagreement lives in values Cassandra cannot produce, which is why it
was arguable in circles: the deciding evidence must be the RULE plus what our
writer does, never an observation.

29 tests pass (11 golden + 15 unit + 3 new); file-size PASS; fmt clean.
